Consolidate tooltip styling into component
