+++ client/app/queue/CaseTitle.jsx
@@ -11,6 +11,8 @@ import { CATEGORIES } from './constants';
 import { COLORS } from '../constants/AppConstants';
 import ReaderLink from './ReaderLink';
 import { ClipboardIcon } from '../components/RenderFunctions';
+import Tooltip from '../components/Tooltip';
+import COPY from '../../COPY.json';
 
 import { toggleVeteranCaseList } from './uiReducer/uiActions';
 
@@ -78,13 +80,14 @@ class CaseTitle extends React.PureComponent {
       <React.Fragment>
         Veteran ID:&nbsp;
         <CopyToClipboard text={appeal.attributes.vbms_id}>
-          <button type="submit"
-            title="Click to copy Veteran ID"
-            className="cf-apppeal-id"
-            {...clipboardButtonStyling} >
-            {appeal.attributes.vbms_id}&nbsp;
-            <ClipboardIcon />
-          </button>
+          <Tooltip text={COPY.CASE_TITLE_VETERAN_ID_BUTTON_TOOLTIP} position="bottom">
+            <button type="submit"
+              className="cf-apppeal-id"
+              {...clipboardButtonStyling} >
+              {appeal.attributes.vbms_id}&nbsp;
+              <ClipboardIcon />
+            </button>
+          </Tooltip>
         </CopyToClipboard>
       </React.Fragment>
