+++ client/app/queue/components/NewFile.jsx
@@ -1,15 +1,14 @@
 // @flow
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { css } from 'glamor';
-import ReactTooltip from 'react-tooltip';
 import { NewFileIcon } from '../../components/RenderFunctions';
-import { COLORS } from '../../constants/AppConstants';
+import Tooltip from '../../components/Tooltip';
 import { bindActionCreators } from 'redux';
 import { getNewDocuments } from '../QueueActions';
 import type {
   BasicAppeal
 } from '../types/models';
+import COPY from '../../../COPY.json';
 
 type Params = {|
   appeal: BasicAppeal
@@ -22,17 +21,6 @@ type Props = Params & {|
   getNewDocuments: Function
 |};
 
-const tooltipID = 'newfile-tip';
-
-const tooltipStyling = css({
-  [`& > #${tooltipID}`]: {
-    backgroundColor: COLORS.GREY_DARK,
-    padding: '0.5rem 1rem',
-    textAlign: 'center'
-  },
-  [`& > #${tooltipID}:after`]: { display: 'none' }
-});
-
 class NewFile extends React.Component<Props> {
   componentDidMount = () => {
     if (!this.props.docs) {
@@ -42,20 +30,9 @@ class NewFile extends React.Component<Props> {
 
   render = () => {
     if (this.props.docs && this.props.docs.length > 0) {
-      return <React.Fragment>
-        <span data-tip data-for={tooltipID}>
-          <NewFileIcon />
-        </span>
-        <span {...tooltipStyling} >
-          <ReactTooltip
-            effect="solid"
-            id={tooltipID}
-            multiline
-            offset={{ top: '-10px' }} >
-            This case has new <br />documents
-          </ReactTooltip>
-        </span>
-      </React.Fragment>;
+      return <Tooltip id="newfile-tip" text={COPY.NEW_FILE_ICON_TOOLTIP} offset={{ top: '-10px' }}>
+        <NewFileIcon />
+      </Tooltip>;
     }
 
     return null;
